requirements.txt
streamlit
python-barcode
reportlab
Pillow
pandas
openpyxl